tests were fixed endpoints for proteced tafs were added additional fields were added to tag entity checking of adding tag to group was added checking of assigning protected tags was added persmissions were added

+++ src/main/java/org/ohdsi/webapi/tag/TagService.java
@@ -3,8 +3,10 @@ package org.ohdsi.webapi.tag;
 import org.glassfish.jersey.internal.util.Producer;
 import org.ohdsi.webapi.cache.CacheService;
 import org.ohdsi.webapi.service.AbstractDaoService;
+import org.ohdsi.webapi.shiro.Entities.UserEntity;
 import org.ohdsi.webapi.tag.domain.Tag;
 import org.ohdsi.webapi.tag.domain.TagInfo;
+import org.ohdsi.webapi.tag.domain.TagType;
 import org.ohdsi.webapi.tag.dto.TagDTO;
 import org.ohdsi.webapi.tag.repository.TagRepository;
 import org.slf4j.Logger;
@@ -16,6 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
+import javax.ws.rs.ForbiddenException;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
@@ -60,10 +63,19 @@ public class TagService extends AbstractDaoService {
     }
 
     public Tag create(Tag tag) {
+        tag.setType(TagType.CUSTOM);
         List<Integer> groupIds = tag.getGroups().stream()
                 .map(Tag::getId)
                 .collect(Collectors.toList());
         List<Tag> groups = findByIdIn(groupIds);
+        boolean allowCustom = groups.stream()
+                .filter(Tag::isAllowCustom)
+                .count() == groups.size();
+
+        if (!allowCustom) {
+            throw new IllegalArgumentException("Tag can be added only to groups that allows to do it");
+        }
+
         tag.setGroups(new HashSet<>(groups));
         tag.setCreatedBy(getCurrentUser());
         tag.setCreatedDate(new Date());
@@ -106,6 +118,10 @@ public class TagService extends AbstractDaoService {
 
     public TagDTO update(Integer id, TagDTO entity) {
         Tag existing = tagRepository.findOne(id);
+
+        checkOwnerOrAdmin(existing.getCreatedBy().getId());
+        checkType(existing.getType());
+
         Tag toUpdate = this.conversionService.convert(entity, Tag.class);
 
         List<Integer> groupIds = toUpdate.getGroups().stream()
@@ -124,7 +140,25 @@ public class TagService extends AbstractDaoService {
     }
 
     public void delete(Integer id) {
+        Tag existing = tagRepository.findOne(id);
+
+        checkOwnerOrAdmin(existing.getCreatedBy().getId());
+        checkType(existing.getType());
+
         tagRepository.delete(id);
+    }
+
+    private void checkOwnerOrAdmin(Long tagOwnerId) {
+        UserEntity user = getCurrentUser();
+        if (!(user.getId().equals(tagOwnerId) || isAdmin())) {
+            throw new ForbiddenException();
+        }
+    }
+
+    private void checkType(TagType type) {
+        if (!TagType.CUSTOM.equals(type)) {
+            throw new IllegalArgumentException("Only custom tags can be processed");
+        }
     }
 
     private Tag save(Tag tag) {
